+++ packages/pass-style/src/symbol.js
@@ -1,4 +1,4 @@
-import { Fail, q } from '@endo/errors';
+import { Fail, q, hideAndHardenFunction } from '@endo/errors';
 
 const { ownKeys } = Reflect;
 
@@ -38,7 +38,7 @@ harden(isPassableSymbol);
 export const assertPassableSymbol = sym =>
   isPassableSymbol(sym) ||
   Fail`Only registered symbols or well-known symbols are passable: ${q(sym)}`;
-harden(assertPassableSymbol);
+hideAndHardenFunction(assertPassableSymbol);
 
 /**
  * If `sym` is a passable symbol, return a string that uniquely identifies this